--- a/.gitignore
+++ b/.gitignore
@@ -1,8 +1,7 @@
 *.o
 *.a
-*.mlmodel
-*.mlmodelc
 .cache/
+.coreml/
 .test/
 .vs/
 .vscode/
@@ -37,4 +36,6 @@ examples/whisper.objc/whisper.objc.xcodeproj/project.xcworkspace/xcuserdata
 
 extra/bench-gg.txt
 
-*.mlmodel*
+models/*.mlmodel
+models/*.mlmodelc
+models/*.mlpackage