… CMakeLists.txt
… coreml/whisper-encoder-impl.h
… coreml/whisper-encoder-impl.m
… extra/bench-all.sh
… models/convert-whisper-to-coreml.py
… models/download-coreml-model.sh
… models/generate-coreml-interface.sh
… models/generate-coreml-model.sh
… whisper.cpp