### Running Whisper inference on Apple Neural Engine (ANE) via Core ML

This PR extends `whisper.cpp` to run the Whisper Encoder on the ANE through Core ML inference.
The performance gain is more than x3 compared to 8-thread CPU for `tiny`, `base` and `small` models.

Here are initial performance benchmarks for the Encoder on M1 Pro with (***top***) and without (***bottom***) Core ML:

| CPU | OS | Config | Model | Th | Load [ms] | Encode [ms] | Commit |
| --- | -- | ------ | ----- | ------- | --------- | ----------- | ------ |
| MacBook M1 Pro | MacOS 13.2.1 |  CORE ML | tiny | 4 | 50 | 30 | b0ac915 |
| MacBook M1 Pro | MacOS 13.2.1 |  CORE ML | base | 4 | 74 | 64 | b0ac915 |
| MacBook M1 Pro | MacOS 13.2.1 |  CORE ML | small | 4 | 188 | 208 | b0ac915 |
| MacBook M1 Pro | MacOS 13.2.1 |  CORE ML | medium | 4 | 533 | 1033 | b0ac915 |
| MacBook M1 Pro | MacOS 13.2.1 |  CORE ML | large | 4 | ? | ? | b0ac915 |
| --- |
| MacBook M1 Pro | MacOS 13.0.1 |  NEON BLAS | tiny | 8 | 71 | 102 | 206fc93 |
| MacBook M1 Pro | MacOS 13.0.1 |  NEON BLAS | base | 8 | 96 | 220 | 206fc93 |
| MacBook M1 Pro | MacOS 13.0.1 |  NEON BLAS | small | 8 | 233 | 685 | 206fc93 |
| MacBook M1 Pro | MacOS 13.0.1 |  NEON BLAS | medium | 8 | 603 | 1928 | 206fc93 |
| MacBook M1 Pro | MacOS 13.0.1 |  NEON BLAS | large | 8 | 1158 | 3350 | 206fc93 |
| --- |

This PR adds a helper script [models/generate-coreml-model.sh](https://github.com/ggerganov/whisper.cpp/blob/coreml/models/generate-coreml-model.sh) that can be used to easily generate a Core ML Encoder model yourself. For now, I don't plan on hosting the Core ML models as there is some chance that the implementation will change in the future. Therefore, it is recommended that everyone simply generate them locally with that script. See the instructions below.

There are a couple of drawbacks:

- First time running a Core ML on a device takes a long time (several seconds, depending on the model). All other runs are fast
- The `medium` and `large` models take a long time to be converted to Core ML (tens of minutes) and require a lot of RAM. First run on a device is also very slow for them, so not sure if these are viable for production use

### Acknowledgements

Huge thanks to @wangchou for the initial demonstration of how to use Core ML in `whisper.cpp` (#548)

Thanks to @RobertRiachi for optimizing for ANE execution and improving the model export process

Thanks to everyone else who participated in #548 and helped with insights, testing and ideas

### Usage

- Install dependencies:

  ```bash
  pip install ane_transformers
  pip install openai-whisper
  pip install coremltools
  ```

- Generate a Core ML model. For example, to generate a `base.en` model, use:

  ```bash
  ./models/generate-coreml-model.sh base.en
  ```

  This will generate the folder `models/ggml-base.en-encoder.mlmodelc`

- Build `whisper.cpp` with Core ML support:

  ```bash
  # using Makefile
  make clean
  WHISPER_COREML=1 make -j
  
  # using CMake
  cd build
  cmake -DWHISPER_COREML=1 ..
  ```

- Run the examples as usual. For example:

  ```bash
  ./main -m models/ggml-base.en.bin -f samples/gb0.wav

  ...

  whisper_init_state: loading Core ML model from 'models/ggml-base.en-encoder.mlmodelc'
  whisper_init_state: first run on a device may take a while ...
  whisper_init_state: Core ML model loaded

  system_info: n_threads = 4 / 10 | AVX = 0 | AVX2 = 0 | AVX512 = 0 | FMA = 0 | NEON = 1 | ARM_FMA = 1 | F16C = 0 | FP16_VA = 1 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 0 | VSX = 0 | COREML = 1 | 

  ...
  ```

  The first run on a device is slow, since the ANE service compiles the Core ML model to some device-specific format.
  Next runs are faster.

### TODO

- [x] Can the Decoder be ported to ANE too? https://github.com/ggerganov/whisper.cpp/discussions/548#discussioncomment-5199310
  **Answer: No**
- [x] Convert the `medium` and `large` models to Core ML format and upload to HF
  Need a Mac Silicon with 64GB RAM to do the conversion from PyTorch -> Core ML
  **Does not seem viable - too slow**
- [x] Unified `ggml` + `coreml` model file
  We currently load both the full `ggml` model (encoder + decoder) and the `coreml` encoder - not optimal
  **Will be done in the future, hopefully via community contributions**
- [x] Scripts for generating Core ML model files (e.g. https://github.com/wangchou/callCoreMLFromCpp)
- [x] Support loading Core ML model from memory buffer
  Currently we support only loading from a folder on the disk
  **Low-prio, hoping for contributions**
- [x] Progress report for initial-run model processing
  **Does not look possible. Any CoreML experts?**
- [x] Adjust memory usage buffers when using Core ML
  **Not needed - the Encoder compute buffer is less than 20MB even for the `large` model**
- [x] Try to avoid the first on-device automatic model generation (it takes a long time)
  **Does not look possible. Any CoreML experts?**
- [x] The `medium` model takes more than 30 minutes to convert on the first run. Is there a work-around?
  **I think no**
- [x] Can we run the Core ML inference on the GPU?
  **Looks like not worth it**

### Future work

- Fix the ANE-optimized Whisper imeplementation. Currently, there is something wrong with the tensor shapes when passed to from / to `whisper.cpp` and the transcription gets corrupted. The optimized version should be about 1.5x faster than the original one
- Add support for decoder-only `ggml` models. This will avoid having to store the Encoder data 2 times on disk / memory. Currently, it is store one time in the `ggml` model and another time in the Core ML model. This will reduce both disk and memory usage
- Add support for running the Decoder on the ANE. Due to the nature of the Decoder operations, it seems that running on the CPU is generally more efficient in terms of speed compared to running it on the ANE. However, an ANE Decoder should be much more energy-efficient compared to the CPU one, so having this option could be useful in some cases