@brozkrut 2023-03-06
Great work!

I tested coreml branch on Mac Mini M2 (base $599 model).

The performance gain seems to be more than **x5** compared to 4-thread CPU (thanks to much faster ANE on M2, 8-thread CPU on Mac Mini M2 base model is slower than 4-thread).

Performance benchmarks for the Encoder with _(top)_ and without _(bottom)_ Core ML:
| CPU | OS | Config | Model | Th | Load | Enc. | Commit |
| --- | -- | ------ | ----- | -- | ---- | ---- | ------ |
| Mac Mini M2 | macOS 13.2.1 |  CORE ML | tiny | 4 | 44 | 25 | 17a1459 |
| Mac Mini M2 | macOS 13.2.1 |  CORE ML | base | 4 | 66 | 54 | 17a1459 |
| Mac Mini M2 | macOS 13.2.1 |  CORE ML | small | 4 | 163 | 190 | 17a1459 |
| Mac Mini M2 | macOS 13.2.1 |  CORE ML | medium | 4 |  |  | 17a1459 |
| Mac Mini M2 | macOS 13.2.1 |  CORE ML | large | 4 |  |  | 17a1459 |
---
| CPU | OS | Config | Model | Th | Load | Enc. | Commit |
| --- | -- | ------ | ----- | -- | ---- | ---- | ------ |
| Mac Mini M2 | macOS 13.2.1 |  NEON BLAS | tiny | 4 | 40 | 142 | 59fdcd1 |
| Mac Mini M2 | macOS 13.2.1 |  NEON BLAS | base | 4 | 67 | 299 | 59fdcd1 |
| Mac Mini M2 | macOS 13.2.1 |  NEON BLAS | small | 4 | 152 | 980 | 59fdcd1 |
| Mac Mini M2 | macOS 13.2.1 |  NEON BLAS | medium | 4 |  |  | 59fdcd1 |
| Mac Mini M2 | macOS 13.2.1 |  NEON BLAS | large | 4 |  |  | 59fdcd1 |
---