@DontEatOreo 2023-03-09
I compiled whisper.cpp with coreml support using `make` as well I build the mlmodel but I'm getting an error
```
whisper_init_from_file: loading model from 'models/ggml-base.en.mlmodelc'
whisper_model_load: loading model
whisper_model_load: invalid model data (bad magic)
whisper_init: failed to load model
error: failed to initialize whisper context
```

Is there anything else I'm missing? 🤔