@ggerganov 2023-03-09
@DontEatOreo 

On the command line, you still have to specify the non-coreml model: `models/ggml-base.en.bin`.
The code will automatically also load the `models/ggml-base.en.mlmodelc` if it is present in the same folder.