@DontEatOreo 2023-03-09
@ggerganov Благодаря ти! I was very confused why it wasn't working even though I did everything right