@strangelearning 2023-03-24
Just saw this was announced, is it useful? https://github.com/apple/ml-ane-transformers