@cerupcat 2023-04-05
> @DontEatOreo
> 
> On the command line, you still have to specify the non-coreml model: `models/ggml-base.en.bin`. The code will automatically also load the `models/ggml-base.en.mlmodelc` if it is present in the same folder.

Does this mean we have to bundle both files with the app? Asking since the file size gets fairly large having to include them all.