@ggerganov 2023-04-14
> > @DontEatOreo
> > On the command line, you still have to specify the non-coreml model: `models/ggml-base.en.bin`. The code will automatically also load the `models/ggml-base.en.mlmodelc` if it is present in the same folder.
> 
> Does this mean we have to bundle both files with the app? Asking since the file size gets fairly large having to include them all.

The solution is to produce encoder-only CoreML model in one file and decoder-only standard model in another file. This is not very difficult to achieve, but supporting so many model files might get too difficult for me. So probably I will rely on someone helping out and demonstrating how this can be done, either as an example in this repo or in a fork.