@ggerganov 2023-04-14
This is getting almost ready to merge. I am hoping to do it tomorrow.

The most important part that currently needs testing is the creation of the CoreML models, following the instructions here:

https://github.com/ggerganov/whisper.cpp/discussions/548#discussioncomment-5327027

If you give this a try, please let us know the results and if you encountered any issues.
Also, lets us know if you used quantized or not-quantized CoreML models and what has been the experience.

I believe that `tiny`, `base` and `small` models should be supported, while `medium` and `large` seem to not be viable for this approach.