@aehlke 2023-04-14
1.4gb for medium sounds fine for users, but you're saying there are other limitations against it?