@neurostar 2023-04-15
| CPU | OS | Config | Model | Th | Load | Enc. | Commit |
| --- | -- | ------ | ----- | -- | ---- | ---- | ------ |
| MacBook Air M2 | MacOS 13.3.1 |  NEON BLAS COREML | tiny | 4 | 41 | 31 | f19e23f |
| MacBook Air M2 | MacOS 13.3.1 |  NEON BLAS COREML | base | 4 | 59 | 57 | f19e23f |
| MacBook Air M2 | MacOS 13.3.1 |  NEON BLAS COREML | small | 4 | 147 | 195 | f19e23f |
| MacBook Air M2 | MacOS 13.3.1 |  NEON BLAS COREML | medium | 4 | 576 | 783 | f19e23f |
| MacBook Air M2 | MacOS 13.3.1 |  NEON BLAS COREML | large | 4 | 1196 | 2551 | f19e23f |

Great work!
It was consuming ~10GB memory converting to ML model, it worked fine on 8GB Air.
It was also taking only 2-3 min for initial loading of large model.