@CarberryChai 2023-04-16
When running this script:
```shell
./models/generate-coreml-model.sh base.en
```
I got the error:
```
xcrun: error: unable to find utility "coremlc", not a developer tool or in PATH
```