@ggerganov 2023-04-16
For now, you should generate the Core ML models locally following the instructions.
I don't want to host them on HF yet, because it is very likely that the models will change soon - there are a some pending improvements (see https://github.com/ggerganov/whisper.cpp/discussions/548#discussioncomment-5622733). If I upload them now, later we will get new models and everyone will be confused which model they are using, etc.