@MathanGeurtsen 2023-04-20
Hey, for people who are still struggling with trying this out, here is a bit of a script that shows what my environment is like. I've had a bit of trouble getting this running, so here is my attempt to make a minimally reproducible set of commands. I haven't checked speed improvements yet. This is on a macbook air M1, running Ventura 13.3.1


```
# I also had problems with xcode, this helped me when it didn't work after reinstalling xcode, thanks @neurostar 
sudo xcode-select --switch /Applications/Xcode.app/Contents/Developer

# setting up coreml conda environment, 3.9 works for me
conda create --name core_whisper_3_9 python=3.9 -y 
conda activate core_whisper_3_9 
# torchvision and tensorflow aren't necessary, but I've had a bit of a problem getting coreml itself to work, so this is to run their example program
pip install torchvision==0.15.1 tensorflow-macos==2.9 coremltools==6.3 ane-transformers==0.1.1 openai-whisper==20230314

# checking whether coreml itself works correctly, no whisper.cpp involvement, code copied from [coremltools.readme.io](https://coremltools.readme.io/docs/convert-a-torchvision-model-from-pytorch)
python -c 'import coremltools as ct
import torch
import torchvision
# Load PyTorch model (and perform tracing)
torch_model = torchvision.models.mobilenet_v2()
torch_model.eval() 

example_input = torch.rand(1, 3, 256, 256)
traced_model = torch.jit.trace(torch_model, example_input)



# Convert using the same API. Note that we need to provide "inputs" for pytorch conversion.
model_from_torch = ct.convert(traced_model,
                              inputs=[ct.TensorType(name="input", 
                                                    shape=example_input.shape)],
                              debug=True)'


# ensure whisper.cpp repo is in the same state as mine was
cd <mypath>/whisper.cpp
git pull
git checkout v1.3.0
git clean -idx # interactive: remove downloaded models. you can skip this step, but it ensures you have a fresh install of the models

# building whisper.cpp with coreml support
mkdir build
cd build
cmake -DWHISPER_COREML=1 ..
make -j
cd ..

# download model and convert to coreml
bash ./models/download-ggml-model.sh tiny
bash ./models/generate-coreml-model.sh tiny

# try out jfk sample
./build/bin/main -m models/ggml-tiny.bin -f samples/jfk.wav
```

This led to the following output for the sample for me:
```
whisper_init_from_file_no_state: loading model from 'models/ggml-tiny.bin'
whisper_model_load: loading model
whisper_model_load: n_vocab       = 51865
whisper_model_load: n_audio_ctx   = 1500
whisper_model_load: n_audio_state = 384
whisper_model_load: n_audio_head  = 6
whisper_model_load: n_audio_layer = 4
whisper_model_load: n_text_ctx    = 448
whisper_model_load: n_text_state  = 384
whisper_model_load: n_text_head   = 6
whisper_model_load: n_text_layer  = 4
whisper_model_load: n_mels        = 80
whisper_model_load: f16           = 1
whisper_model_load: type          = 1
whisper_model_load: mem required  =  129.00 MB (+    3.00 MB per decoder)
whisper_model_load: adding 1608 extra tokens
whisper_model_load: model ctx     =   73.58 MB
whisper_model_load: model size    =   73.54 MB
whisper_init_state: kv self size  =    2.62 MB
whisper_init_state: kv cross size =    8.79 MB
whisper_init_state: loading Core ML model from 'models/ggml-tiny-encoder.mlmodelc'
whisper_init_state: first run on a device may take a while ...
whisper_init_state: Core ML model loaded

system_info: n_threads = 4 / 8 | AVX = 0 | AVX2 = 0 | AVX512 = 0 | FMA = 0 | NEON = 1 | ARM_FMA = 1 | F16C = 0 | FP16_VA = 1 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 0 | VSX = 0 | COREML = 1 |

main: processing 'samples/jfk.wav' (176000 samples, 11.0 sec), 4 threads, 1 processors, lang = en, task = transcribe, timestamps = 1 ...


[00:00:00.000 --> 00:00:10.500]   And so my fellow Americans ask not what your country can do for you ask what you can do for your country.


whisper_print_timings:     load time =    97.74 ms
whisper_print_timings:     fallbacks =   0 p /   0 h
whisper_print_timings:      mel time =    41.69 ms
whisper_print_timings:   sample time =    10.67 ms /    25 runs (    0.43 ms per run)
whisper_print_timings:   encode time =    33.32 ms /     1 runs (   33.32 ms per run)
whisper_print_timings:   decode time =    38.86 ms /    25 runs (    1.55 ms per run)
whisper_print_timings:    total time =  4314.09 ms
```

Hope this helps someone.