@dreampuf 2023-06-21
After couple times of restarting the main process, CoreML works. 

```
# without coreml
whisper_print_timings:     load time =  1552.45 ms
whisper_print_timings:     fallbacks =   0 p /   0 h
whisper_print_timings:      mel time =   747.98 ms
whisper_print_timings:   sample time =  1122.64 ms /  2705 runs (    0.42 ms per run)
whisper_print_timings:   encode time = 51066.02 ms /    22 runs ( 2321.18 ms per run)
whisper_print_timings:   decode time = 62040.41 ms /  2703 runs (   22.95 ms per run)
whisper_print_timings:    total time = 116702.45 ms
```

```
# with coreml
whisper_print_timings:     load time =   964.80 ms
whisper_print_timings:     fallbacks =   0 p /   0 h
whisper_print_timings:      mel time =   801.19 ms
whisper_print_timings:   sample time =  1143.31 ms /  2711 runs (    0.42 ms per run)
whisper_print_timings:   encode time = 24224.26 ms /    22 runs ( 1101.10 ms per run)
whisper_print_timings:   decode time = 64989.42 ms /  2709 runs (   23.99 ms per run)
whisper_print_timings:    total time = 93745.79 ms
```